+++ src/app/pager/pager-standalone.demo.ts
@@ -35,6 +35,28 @@ export class PagerStandaloneDemoComponent {
     pageSizes: [5, 10, 15, 20],
   };
 
+  modelNew = {
+    dataset: PAGING_DATA.slice(80),
+    hideFirstButton: false,
+    hidePreviousButton: false,
+    hideNextButton: false,
+    hideLastButton: false,
+    disableFirstButton: false,
+    disableLastButton: false,
+    disablePreviousButton: false,
+    disableNextButton: false,
+    firstPageTooltip: 'click to got to the first page of records',
+    lastPageTooltip: 'click to got to the last page of records',
+    previousPageTooltip: 'click to got to the previous page of records',
+    nextPageTooltip: 'click to got to the last page of records',
+    hidePageSelectorInput: false,
+    hidePageSizeSelector: false,
+    useSmallPageSizeSelector: false,
+    pageSizeMenuSettings: { attachToBody: false },
+    pageSize: 10,
+    pageSizes: [5, 10, 15, 20],
+  };
+
   showModel = false;
 
   onFirstPage(_: any) {
@@ -59,5 +81,9 @@ export class PagerStandaloneDemoComponent {
 
   toggleModel() {
     this.showModel = !this.showModel;
+  }
+
+  changeDataset() {
+    this.model = this.modelNew;
   }
 }