+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 17.3.0 Fixes
 
 - `[BusyIndicator]` Added optional settings in the updated method. ([#1602](https://github.com/infor-design/enterprise-ng/issues/1602))
+- `[Pager]` Fixed dataset not updated correctly. ([#1554](https://github.com/infor-design/enterprise-ng/issues/1554))
 
 ## 17.1.0